DeepSeek_Simple_Bot/requirements.txt
aiogram>=3.0,<4
openai>=1.0
python-dotenv>=1.0
